same user delete check

+++ service/user_http.go
@@ -247,11 +247,21 @@ func deleteUserHandler(deps Dependencies) http.HandlerFunc {
 
 		vars := mux.Vars(req)
 		username := vars["username"]
+
+		currentUser := req.Context().Value(db.ContextKeyUsername).(string)
+		if currentUser == username {
+			logger.WithField("err", "SAMEUSERERROR").Error(responses.SameUserDeleteError)
+			responseCodeAndMsg(rw, http.StatusForbidden, ErrObj{Err: responses.SameUserDeleteError.Error()})
+			return
+		}
 		err := deps.Store.DeleteUser(req.Context(), username)
 		if err != nil {
 			logger.WithField("err", err.Error())
 			responseCodeAndMsg(rw, http.StatusInternalServerError, ErrObj{Err: responses.UserDeleteError.Error()})
+			return
 		}
 
+		logger.Infoln(responses.UserDeleteSuccess)
+		responseCodeAndMsg(rw, http.StatusOK, MsgObj{Msg: responses.UserDeleteSuccess})
 	})
 }
